+++ .agentplane/tasks/202605290344-YV53ZN/pr/github-body.md
@@ -1,0 +1,33 @@
+Task: `202605290344-YV53ZN`
+Title: PR review template hosted formatting decomposition
+Canonical task record: `.agentplane/tasks/202605290344-YV53ZN/README.md`
+
+## Summary
+
+PR review template hosted formatting decomposition
+
+Decompose packages/agentplane/src/commands/pr/internal/review-template.ts by extracting hosted GitHub verification markdown command-wrapping helpers into a focused module, preserving PR artifact rendering and validation behavior while reducing hotspot count.
+
+## Scope
+
+- In scope: Decompose packages/agentplane/src/commands/pr/internal/review-template.ts by extracting hosted GitHub verification markdown command-wrapping helpers into a focused module, preserving PR artifact rendering and validation behavior while reducing hotspot count.
+- Out of scope: unrelated refactors not required for "PR review template hosted formatting decomposition".
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T03:44:48.756Z
+- Branch: task/202605290344-YV53ZN/pr-review-template-hosted-formatting
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
