+++ .agentplane/tasks/202605290344-YV53ZN/pr/review.md
@@ -1,0 +1,36 @@
+# PR Review
+
+Created: 2026-05-29T03:44:48.756Z
+
+## Task
+
+- Task: `202605290344-YV53ZN`
+- Title: PR review template hosted formatting decomposition
+- Status: DOING
+- Branch: `task/202605290344-YV53ZN/pr-review-template-hosted-formatting`
+- Canonical task record: `.agentplane/tasks/202605290344-YV53ZN/README.md`
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T03:44:48.756Z
+- Branch: task/202605290344-YV53ZN/pr-review-template-hosted-formatting
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
